feat: check if the bin is compiled before trying to run it exec_rss_notify.sh will now to a check to see if the program is compiled to the expected location before trying to run it. If not, tell the user to compile the program, then exit.

+++ exec_rss_notify.sh
@@ -1,13 +1,19 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-echo "Launching rss_notify on $(date +%Y-%m-%d-%H:%M%S)"
+echo "Launching rss notify on $(date +%Y-%m-%d-%H:%M%S)"
 
 echo "Sourcing env vars."
 SCRIPT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
 source "${SCRIPT_DIR}"/.env
 
-pkill -f "${RSS_NOTIFY_BIN}" || true
+if [[ ! -f "${RSS_NOTIFY_BIN}" ]]; then
+	echo "rss notify is not compiled. Try running \"cargo build --release\" and ensure the bin dir env variable is set to the proper location for your binary."
+	exit 1
+else
+	echo "Killing previous instance of rss notify if it is currently running."
+	pkill -f "${RSS_NOTIFY_BIN}" || true
+fi
 
 RSS_NOTIFY_LOG_DIR="$(dirname "${RSS_NOTIFY_LOG_FILE}")"
 
@@ -15,6 +21,7 @@ if [[ ! -d "${RSS_NOTIFY_LOG_DIR}" ]]; then
 	echo "Logging directory for rss_notify ${RSS_NOTIFY_LOG_DIR} does not exist, making it now."
 	mkdir -p "${RSS_NOTIFY_LOG_DIR}"
 else
+	echo "Performing log cleanup."
 	find "${RSS_NOTIFY_LOG_DIR}" -type f -mtime +"${RSS_NOTIFY_LOG_RETENTION_DAYS}" -delete
 fi
